refactor: finish global state dispatch cleanup
