+++ src/global_state/dispatch.rs
@@ -1,84 +1,2 @@
-use lsp_server::Request;
-use lsp_types::request::Request as _;
-
-use super::{GlobalState, dispatcher::ReqDispatcher, handlers};
-use crate::i18n::keys;
-
-impl GlobalState {
-    pub(in crate::global_state) fn handle_request(&mut self, req: Request) {
-        if Self::is_pull_diagnostic_request(&req) && !self.is_workspace_ready() {
-            self.workspace.workspace_vfs.defer_diagnostics_until_ready();
-            self.diagnostics.pending_diagnostic_requests.push(req);
-            return;
-        }
-
-        let mut dispatcher = ReqDispatcher { req: Some(req), global_state: self };
-
-        // Handle shutdown req first
-        dispatcher.on_sync_mut::<lsp_types::request::Shutdown>(|this, ()| {
-            this.client.shutdown_requested = true;
-            this.cancel_all_tasks();
-            Ok(())
-        });
-
-        match &mut dispatcher {
-            ReqDispatcher { req: Some(req), global_state: this }
-                if this.client.shutdown_requested =>
-            {
-                this.respond(lsp_server::Response::new_err(
-                    req.id.clone(),
-                    lsp_server::ErrorCode::InvalidRequest as i32,
-                    this.config_state
-                        .config
-                        .i18n
-                        .text(keys::SERVER_SHUTDOWN_ALREADY_REQUESTED)
-                        .to_owned(),
-                ));
-                return;
-            }
-            _ => (),
-        }
-
-        use handlers::request::*;
-        use lsp_types::request::*;
-        dispatcher
-            .on_no_retry::<Completion>(handle_completion)
-            .on_latency_sensitive::<SemanticTokensFullRequest>(handle_semantic_tokens_full)
-            .on_latency_sensitive::<SemanticTokensFullDeltaRequest>(
-                handle_semantic_tokens_full_delta,
-            )
-            .on_latency_sensitive::<SemanticTokensRangeRequest>(handle_semantic_tokens_range)
-            .on::<DocumentSymbolRequest>(handle_document_symbol)
-            .on::<WorkspaceSymbolRequest>(handle_workspace_symbol)
-            .on::<FoldingRangeRequest>(handle_folding_ranges)
-            .on::<DocumentDiagnosticRequest>(handle_document_diagnostic)
-            .on::<WorkspaceDiagnosticRequest>(handle_workspace_diagnostic)
-            .on_no_retry::<SignatureHelpRequest>(handle_signature_help)
-            .on_no_retry::<InlayHintRequest>(handle_inlay_hint)
-            .on_no_retry::<CodeLensRequest>(handle_code_lens)
-            .on_no_retry::<CodeLensResolve>(handle_code_lens_resolve)
-            .on_no_retry::<HoverRequest>(handle_hover)
-            .on_no_retry::<GotoDefinition>(handle_goto_definition)
-            .on_no_retry::<GotoDeclaration>(handle_goto_declaration)
-            .on_no_retry::<DocumentHighlightRequest>(handle_document_highlight)
-            .on_no_retry::<References>(handle_references)
-            .on_no_retry::<PrepareRenameRequest>(handle_prepare_rename)
-            .on_no_retry::<Rename>(handle_rename)
-            .on_fmt_thread::<Formatting>(handle_formatting)
-            .on_fmt_thread::<RangeFormatting>(handle_range_formatting)
-            .on_fmt_thread::<OnTypeFormatting>(handle_on_type_formatting)
-            .on_no_retry::<CodeActionRequest>(handle_code_action)
-            .on_no_retry::<CodeActionResolveRequest>(handle_code_action_resolve)
-            .on_sync_mut::<ExecuteCommand>(handle_execute_command)
-            .on::<SelectionRangeRequest>(handle_selection_range)
-            .finish();
-    }
-
-    fn is_pull_diagnostic_request(req: &Request) -> bool {
-        matches!(
-            req.method.as_str(),
-            lsp_types::request::DocumentDiagnosticRequest::METHOD
-                | lsp_types::request::WorkspaceDiagnosticRequest::METHOD
-        )
-    }
-}
+mod notification;
+mod request;
